requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.108.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
pinterest-api-sdk==0.2.1
Pinterest-Generated-Client==0.1.7
pluggy==1.3.0
pydantic==2.5.3
pydantic-core==2.14.6
pydantic-settings==2.1.0
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
tomli==2.0.1
typing-extensions==4.9.0
urllib3==1.26.12
uvicorn==0.25.0
